+++ src/onInit/summarizeData/calculateStatistics.js
@@ -57,7 +57,8 @@ export default function calculateStatistics(onInit = true) {
             });
             summary.nest_level = d[0].nest_level;
             summary.parents = d[0].parents;
-            summary.folder_ordinal = d[0].folder_ordinal;
+            summary.visit_order = d[0][context.initial_config.visit_order_col];
+            summary.form_order = d[0][context.initial_config.form_order_col];
             return summary;
         })
         .entries(this.data.initial_filtered);
@@ -71,7 +72,8 @@ export default function calculateStatistics(onInit = true) {
         });
         d.nest_level = d.values.nest_level;
         d.parents = d.values.parents;
-        d.folder_ordinal = d.values.folder_ordinal;
+        d.visit_order = d.values.visit_order;
+        d.form_order = d.values.form_order;
 
         delete d.values;
     });
